refactor: extract materialization merge
