+++ apps/sva-studio-react/src/lib/waste-management-mainserver-sync.materialization.ts
@@ -15,6 +15,7 @@ import {
   type MaterializationRule,
   type WasteMaterializationContext,
 } from './waste-management-mainserver-sync.materialization.shared.js';
+import { mergeMaterializedPickupDates } from './waste-management-mainserver-sync.materialization.merge.js';
 
 const isRuleApplicableToTour = (rule: MaterializationRule, tourId: string): boolean => {
   if (rule.appliesToTourId && rule.appliesToTourId !== tourId) {
@@ -228,34 +229,6 @@ export type MaterializedLocationTourPickupDateRecord = Omit<
     readonly createdAt: string;
     readonly updatedAt: string;
   }>;
-
-const toOccurrenceKey = (
-  entry: Pick<MaterializedLocationTourPickupDateRecord, 'locationId' | 'tourId' | 'pickupDate'>
-): string => `${entry.locationId}::${entry.tourId}::${entry.pickupDate}`;
-
-const mergeMaterializedPickupDates = (
-  calculated: readonly MaterializedLocationTourPickupDateRecord[],
-  assigned: readonly MaterializedLocationTourPickupDateRecord[]
-): readonly MaterializedLocationTourPickupDateRecord[] => {
-  const assignedKeys = new Set(assigned.map(toOccurrenceKey));
-  const seenCalculatedKeys = new Set<string>();
-  const uniqueCalculated = calculated.filter((entry) => {
-    const key = toOccurrenceKey(entry);
-    if (assignedKeys.has(key) || seenCalculatedKeys.has(key)) {
-      return false;
-    }
-    seenCalculatedKeys.add(key);
-    return true;
-  });
-
-  return [...uniqueCalculated, ...assigned].sort(
-    (left, right) =>
-      left.locationId.localeCompare(right.locationId) ||
-      left.tourId.localeCompare(right.tourId) ||
-      left.pickupDate.localeCompare(right.pickupDate) ||
-      left.id.localeCompare(right.id)
-  );
-};
 
 const materializeLinkedPickupDates = (input: {
   readonly link: WasteMaterializationContext['links'][number];
